Add custom score range option for relevance judgments
- Add --score-scale flag: 'trec' (0-3) or 'range' (custom min/max)
- Add --score-min and --score-max flags for custom range bounds
- Update Qrel to use u16 for larger score ranges
- Add range_relevance_prompt with dynamic thresholds
- Add score_range parsing and validation in LLM provider
- Update RelevanceScorer with with_scale constructor
- Update README with score scale documentation

+++ src/output/beir.rs
@@ -26,7 +26,7 @@ pub struct BeirQuery {
 pub struct Qrel {
     pub query_id: String,
     pub doc_id: String,
-    pub score: u8,
+    pub score: u16,
 }
 
 /// Write documents to a BEIR corpus.jsonl file
